Added mutation testing dependency
NOTE: We need a pitest junit plugin for pitest plugin to work.
This is why we add the dependency - pitest-junit5-plugin - in pom.xml
<plugin>
    <groupId>org.pitest</groupId>
    <artifactId>pitest-maven</artifactId>
    <version>1.15.0</version>
    <dependencies>
        <dependency>
            <groupId>org.pitest</groupId>
            <artifactId>pitest-junit5-plugin</artifactId>
            <version>1.2.0</version>
        </dependency>
    </dependencies>
</plugin>

To run the mutation tests, we have to make sure that our test
run is successful. Only then can we run the task/goal -
pitest:mutationCoverage either via UI or CLI like so -
mvn test-compile org.pitest:pitest-maven:mutationCoverage